event exceptions

+++ WEB-INF/classes/servlet/DeleteEventServlet.java
@@ -14,7 +14,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.transaction.UserTransaction;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.time.LocalDate;
+import java.util.Objects;
 
 import static java.lang.System.out;
 
@@ -40,25 +42,34 @@ public class DeleteEventServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         // deleting an event from the database
+        response.setContentType("text/html");
+        PrintWriter out = response.getWriter();
+        out.println("<HTML>");
+
         String id = request.getParameter("id_del");
-        int idInt = Integer.parseInt(id);
-        try {
-            Event event = em.find(Event.class, idInt );
-            if (event == null) {
-                out.println("Event not found");
-            } else {
-                ut.begin();
-                if (!em.contains(event)) {
-                    event = em.merge(event);
+        if (Objects.equals(id, "")) {
+            out.println("<h2>Please, enter an event id.</h2>");
+        }else {
+            int idInt = Integer.parseInt(id);
+            try {
+                Event event = em.find(Event.class, idInt);
+                if (event == null) {
+                    out.println("Event not found");
+                } else {
+                    ut.begin();
+                    if (!em.contains(event)) {
+                        event = em.merge(event);
+                    }
+                    em.remove(event);
+                    ut.commit();
+                    out.println("<script>window.location.href='loggedAdmin.html';</script>");
                 }
-                em.remove(event);
-                ut.commit();
-                out.println("<script>window.location.href='loggedAdmin.html';</script>");
+
+            } catch (Exception e) {
+                out.println("Error");
             }
-
-        } catch (Exception e) {
-            out.println("Error");
         }
-
+        out.println("</HTML>");
+        out.close();
     }
 }
